+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # Okta Node SDK Changelog
 
+# 7.2.1
+
+### Fixes
+
+- [#442](https://github.com/okta/okta-sdk-nodejs/pull/442)
+  - Fixed schemas:
+    - OrgCAPTCHASettings
+    - EmailServerListResponse
+    - UISchemaObject (elements)
+  - Fixed return type for:
+    - applicationApi.listJwk
+    - logStreamApi.activateLogStream
+    - logStreamApi.deactivateLogStream
+    - emailServerApi.getEmailServer
+  - Fixed request type for:
+    - logStreamApi.createLogStream
+  - Fixes detection of affected resources for caching middleware
+
 # 7.2.0
 
 ### Features